Fix 2, refactor CheckoutPage to several classes, change FirstPage->ProductDetailsPage
