+++ src/test/java/automationpractice/com/features/checkout/AddAddressesFeatures.java
@@ -1,7 +1,9 @@
 package automationpractice.com.features.checkout;
 
 import automationpractice.com.data.NewAddressInformation;
+import automationpractice.com.pages.AddressCheckoutPage;
 import automationpractice.com.pages.CheckoutPage;
+import automationpractice.com.pages.SummaryCheckoutPage;
 import automationpractice.com.steps.LoginSteps;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
@@ -14,29 +16,30 @@ public class AddAddressesFeatures {
 
     private CheckoutPage checkoutPage;
 
-    @Steps
-    private LoginSteps loginSteps;
-//    never used
+    private SummaryCheckoutPage summaryCheckoutPage;
+    private AddressCheckoutPage addressCheckoutPage;
+
+
 
 
     @When("^user input new address$")
     public void userInputNewAddress() {
-        checkoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
-        checkoutPage.clickAddANewAddressButton();
-        checkoutPage.inputExampleAddressesData();
+        summaryCheckoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
+        addressCheckoutPage.clickAddANewAddressButton();
+        addressCheckoutPage.inputExampleAddressesData();
 //        what data? you can't store andy data in page class
     }
 
     @When("^user choices new address as billing address$")
     public void userChoicesNewAddressAsBillingAddress() {
-        checkoutPage.selectBillingAddressFromList();
+        addressCheckoutPage.selectBillingAddressFromList();
     }
 
     @Then("^the new address is selected as billing address$")
     public void theNewAddressIsSelectedAsBillingAddress() {
         Assert.assertEquals((NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getFirstNameAddress()
                         + " " + NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getLastNameAddress())
-                , checkoutPage.isBillingAddressCorrectNameAndLastName());
+                , addressCheckoutPage.isBillingAddressCorrectNameAndLastName());
 /*        If you don't return boolean, method name shouldn't start from is. You can do this by:
        final SoftAssertions softAssertions = new SoftAssertions();
        softAssertions.assertThat(checkoutPage.getBillingAddress()).contains(NEW_ADDRESS_INFORMATION_1.getFirstNameAddress());
@@ -46,12 +49,12 @@ public class AddAddressesFeatures {
         Assert.assertEquals((NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getCity()
                         + ", " + NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getState()
                         + " " + NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getZipCode())
-                , checkoutPage.isBillingAddressCorrectCityAndStateAndZipCode());
+                , addressCheckoutPage.isBillingAddressCorrectCityAndStateAndZipCode());
     }
 
     @When("^user choices new address as delivery and billing address$")
     public void userChoicesNewAddressAsDeliveryAndBillingAddress() {
-        checkoutPage.selectDeliveryAndBillingAddressFromList();
+        addressCheckoutPage.selectDeliveryAndBillingAddressFromList();
     }
 
     @Then("^the new address is selected as delivery and billing address$")
@@ -63,18 +66,18 @@ public class AddAddressesFeatures {
 
     @When("^user choices new address as delivery address$")
     public void userChoicesNewAddressAsDeliveryAddress() {
-        checkoutPage.selectDeliveryAddressFromList();
+        addressCheckoutPage.selectDeliveryAddressFromList();
     }
 
     @Then("^the new address is selected as delivery address$")
     public void theNewAddressIsSelectedAsDeliveryAddress() {
         Assert.assertEquals((NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getFirstNameAddress()
                         + " " + NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getLastNameAddress())
-                , checkoutPage.isDeliveryAddressCorrectNameAndLastName());
+                , addressCheckoutPage.isDeliveryAddressCorrectNameAndLastName());
 
         Assert.assertEquals((NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getCity()
                         + ", " + NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getState()
                         + " " + NewAddressInformation.NEW_ADDRESS_INFORMATION_1.getZipCode())
-                , checkoutPage.isDeliveryAddressCorrectCityAndStateAndZipCode());
+                , addressCheckoutPage.isDeliveryAddressCorrectCityAndStateAndZipCode());
     }
 }
