+++ src/test/java/automationpractice/com/features/checkout/AllPathFromChooseToPay.java
@@ -1,6 +1,6 @@
 package automationpractice.com.features.checkout;
 
-import automationpractice.com.pages.CheckoutPage;
+import automationpractice.com.pages.*;
 import automationpractice.com.steps.LoginSteps;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
@@ -14,32 +14,38 @@ public class AllPathFromChooseToPay {
     @Steps
     private LoginSteps loginSteps;
 
+    private SummaryCheckoutPage summaryCheckoutPage;
+    private AddressCheckoutPage addressCheckoutPage;
+    private ShippingCheckoutPage shippingCheckoutPage;
+    private PaymentCheckoutPage paymentCheckoutPage;
+
+
     @When("^he confirms every step to make order and pay by bank wire$")
     public void heConfirmsEveryStepToMakeOrderAndPayByBankWire() {
-        checkoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
-        checkoutPage.clickOnProceedToCheckoutButtonInAddressLabel();
-        checkoutPage.clickAgreeCheckbox();
-        checkoutPage.clickOnProceedToCheckoutButtonInShippingLabel();
-        checkoutPage.clickPayByBankWireButton();
+        summaryCheckoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
+        addressCheckoutPage.clickOnProceedToCheckoutButtonInAddressLabel();
+        shippingCheckoutPage.clickAgreeCheckbox();
+        shippingCheckoutPage.clickOnProceedToCheckoutButtonInShippingLabel();
+        paymentCheckoutPage.clickPayByBankWireButton();
     }
 
     @When("^he confirms order$")
     public void heConfirmsOrder() {
-        checkoutPage.clickIConfirmMyOrder();
+        paymentCheckoutPage.clickIConfirmMyOrder();
     }
 
     @Then("^order is created$")
     public void orderIsCreated() {
-        Assert.assertTrue(checkoutPage.isOrderConfirmationMessage());
+        Assert.assertTrue(paymentCheckoutPage.isOrderConfirmationMessage());
 //        isOrderConfirmationMessageDisplayed()
     }
 
     @When("^he confirms every step to make order and pay by check$")
     public void heConfirmsEveryStepToMakeOrderAndPayByCheck() {
-        checkoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
-        checkoutPage.clickOnProceedToCheckoutButtonInAddressLabel();
-        checkoutPage.clickAgreeCheckbox();
-        checkoutPage.clickOnProceedToCheckoutButtonInShippingLabel();
-        checkoutPage.clickPayByCheckButton();
+        summaryCheckoutPage.clickOnProceedToCheckoutButtonInSummaryLabel();
+        addressCheckoutPage.clickOnProceedToCheckoutButtonInAddressLabel();
+        shippingCheckoutPage.clickAgreeCheckbox();
+        shippingCheckoutPage.clickOnProceedToCheckoutButtonInShippingLabel();
+        paymentCheckoutPage.clickPayByCheckButton();
     }
 }
