+++ src/test/java/automationpractice/com/features/checkout/EditBeforeAddToCartFeatures.java
@@ -1,67 +1,62 @@
 package automationpractice.com.features.checkout;
 
-import automationpractice.com.pages.CheckoutPage;
-import automationpractice.com.pages.FirstProductPage;
-import automationpractice.com.pages.HomePage;
-import automationpractice.com.steps.LoginSteps;
+import automationpractice.com.pages.*;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import net.thucydides.core.annotations.Steps;
 import org.junit.Assert;
 
 public class EditBeforeAddToCartFeatures {
 
     private HomePage homePage;
-    private FirstProductPage firstProductPage;
+    private ProductDetailsPage productDetailsPage;
     private CheckoutPage checkoutPage;
-    @Steps
-    private LoginSteps loginSteps;
-    //LoginSteps is never used - You can remove it from here
+    private MyAccountPage myAccountPage;
 
+    private SummaryCheckoutPage summaryCheckoutPage;
 
     @When("^user select first item from homepage change quantity used plus and minus button before added to cart$")
     public void userSelectFirstItemFromHomepageChangeQuantityUsedPlusAndMinusButtonBeforeAddedToCart() {
         homePage.clickOnNewProductImage();
-        firstProductPage.clickPlusQuantityButton();
-        firstProductPage.clickPlusQuantityButton();
-        firstProductPage.clickMinusQuantityButton();
-        firstProductPage.clickOnAddToCartButton();
-        firstProductPage.clickProceedToCheckoutButton();
+        productDetailsPage.clickPlusQuantityButton();
+        productDetailsPage.clickPlusQuantityButton();
+        productDetailsPage.clickMinusQuantityButton();
+        productDetailsPage.clickOnAddToCartButton();
+        productDetailsPage.clickProceedToCheckoutButton();
     }
 
     @Then("^checkout page has two items$")
     public void checkoutPageHasTwoItems() {
-        Assert.assertEquals("2", checkoutPage.isQuantityAfterAddToCartEqualsTwo());
+        Assert.assertEquals("2", summaryCheckoutPage.isQuantityAfterAddToCartEqualsTwo());
 //        I commented before
     }
 
     @When("^user select first item from homepage and change size$")
     public void userSelectFirstItemFromHomepageAndChangeSize() {
         homePage.clickOnNewProductImage();
-        firstProductPage.changeSize();
+        productDetailsPage.changeSize();
 //        to with size?
-        firstProductPage.clickOnAddToCartButton();
-        firstProductPage.clickProceedToCheckoutButton();
+        productDetailsPage.clickOnAddToCartButton();
+        productDetailsPage.clickProceedToCheckoutButton();
     }
 
     @Then("^checkout page has item in L size$")
     public void checkoutPageHasItemInLSize() {
-        Assert.assertTrue(checkoutPage.isSizeL());
+        Assert.assertTrue(summaryCheckoutPage.isSizeL());
 //        getSize();
     }
 
     @When("^user select first item from homepage and change color$")
     public void userSelectFirstItemFromHomepageAndChangeColor() {
         homePage.clickOnNewProductImage();
-        firstProductPage.changeColorToBlue();
+        productDetailsPage.changeColorToBlue();
         //changeColor("blue");
-        firstProductPage.clickOnAddToCartButton();
-        firstProductPage.clickProceedToCheckoutButton();
+        productDetailsPage.clickOnAddToCartButton();
+        productDetailsPage.clickProceedToCheckoutButton();
     }
 
     @Then("^checkout page has item in different color$")
     public void checkoutPageHasItemInDifferentColor() {
-        Assert.assertTrue(checkoutPage.isBlueColor());
+        Assert.assertTrue(summaryCheckoutPage.isBlueColor());
 //        getColor();
     }
 }
