+++ src/test/java/automationpractice/com/pages/MyAccountPage.java
@@ -3,7 +3,6 @@ package automationpractice.com.pages;
 import net.serenitybdd.core.pages.PageObject;
 import net.serenitybdd.core.pages.WebElementFacade;
 import org.openqa.selenium.support.FindBy;
-import org.w3c.dom.html.HTMLInputElement;
 
 public class MyAccountPage extends PageObject {
 
@@ -19,4 +18,6 @@ public class MyAccountPage extends PageObject {
     public void clickMyStore() {
         myStore.click();
     }
+
+
 }
