+++ src/test/java/automationpractice/com/pages/ProductDetailsPage.java
@@ -4,7 +4,7 @@ import net.serenitybdd.core.pages.PageObject;
 import net.serenitybdd.core.pages.WebElementFacade;
 import org.openqa.selenium.support.FindBy;
 
-public class FirstProductPage extends PageObject {
+public class ProductDetailsPage extends PageObject {
 
     @FindBy(id = "add_to_cart")
     private WebElementFacade clickOnAddToCartButton;
