+++ src/test/resources/features/socialMedia/shareInformationAboutItem.feature
@@ -5,18 +5,18 @@ Feature: Share information about item
     Given logged customer is on home page
 #    I think here we should be on product details page, not on homepage
 
-  Scenario: User is able to share information about item with friends using twitter
-    When user select first item from homepage and chose twitter
-    Then new page in twitter domain is open
-
-  Scenario: User is able to share information about item with friends using facebook
-    When user select first item from homepage and chose facebook
-    Then new page in facebook domain is open
-
-  Scenario: User is able to share information about item with friends using google plus
-    When user select first item from homepage and chose google plus
-    Then new page in google plus domain is open
-
-  Scenario: User is able to share information about item with friends using pinterest
-    When user select first item from homepage and chose pinterest
-    Then new page in pinterest domain is open
+#  Scenario: User is able to share information about item with friends using twitter
+#    When user select first item from homepage and chose twitter
+#    Then new page in twitter domain is open
+#
+#  Scenario: User is able to share information about item with friends using facebook
+#    When user select first item from homepage and chose facebook
+#    Then new page in facebook domain is open
+#
+#  Scenario: User is able to share information about item with friends using google plus
+#    When user select first item from homepage and chose google plus
+#    Then new page in google plus domain is open
+#
+#  Scenario: User is able to share information about item with friends using pinterest
+#    When user select first item from homepage and chose pinterest
+#    Then new page in pinterest domain is open
